requirements.txt
python-telegram-bot==13.7
async-generator==1.10
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
paramiko==2.7.2
requests